Completed preparation requirements for Anagram's Workshop
Implemented a simple program that uses a Hashmap to map countries' codes to countries' names.
Added code for retrieving a country's code based on the name as well (getting a key based on a value).

Most, useful documentation was commented along the lines.
Still, outstanding documentation for retrieving a key based on a value:
https://stackoverflow.com/questions/1383797/java-hashmap-how-to-get-key-from-value